+++ include/codecs/ProcessorALaw.h
